+++ main/common_utils.h
@@ -11,7 +11,7 @@ void enableUserLED(const bool enableLED);
 
 /// configure wakeup timer
 [[nodiscard("device won't wake on timer if unchecked")]]
-esp_err_t registerWakeupTimer(const uint32_t wakeupMicrosec);
+esp_err_t registerWakeupTimer(const uint64_t wakeupMicrosec);
 esp_err_t registerWakeupTimer(int) = delete("duration must be uint32_t microseconds — negative values silently wrap to enormous sleep");
 
 /// sleep for wakeupMicrosec from above + ?10ms? using esp_light_sleep
